Move StateMachine into a separate file
